refactor:移动类并添加函数到Common.h
移动StaticVariableWrapper类及辅助函数到Common.h中
添加了GetArrayLength用于编译期推断数组长度
