+++ TrafficMonitor/Common.h
@@ -338,3 +338,92 @@ public:
     }
 
 };
+
+/**
+ * @brief 编译期获取数组长度
+ *
+ * @tparam T 数组元素类型
+ * @tparam N 编译期推断的数组长度
+ * @return constexpr std::size_t 编译期推断的数组长度
+ */
+template <typename T, std::size_t N>
+constexpr inline std::size_t GetArrayLength(const T (&)[N]) noexcept
+{
+    return N;
+}
+
+/**
+ * @brief 析构StaticVariableWrapper包装对象前默认执行的函数，实际上无操作
+ *
+ * @tparam T
+ */
+template <class T>
+class DefaultStaticVariableWrapperDtor
+{
+public:
+    void operator()(T*){};
+};
+/**
+ * @brief 设计上用于静态变量包装类，用于自定义变量默认初始化后行为和析构前行为
+ *
+ * @tparam T 要被包装的类型
+ * @tparam DTOR 自定义执行析构函数前的行为
+ */
+template <class T, class DTOR = DefaultStaticVariableWrapperDtor<T>>
+
+class StaticVariableWrapper : private DTOR
+{
+private:
+    T m_content;
+
+public:
+    /**
+     * @brief 构造一个StaticVariableWrapper
+     *
+     * @tparam CTOR 自定义变量默认初始化后的函数类型
+     * @param ctor 自定义变量默认初始化后的行为，传入变量的指针作为参数
+     * @param dtor 自定义变量执行析构函数前的行为，传入变量的指针作为参数
+     */
+    template <class CTOR>
+    StaticVariableWrapper(CTOR ctor, DTOR dtor = {})
+        : DTOR{dtor}
+    {
+        ctor(std::addressof(m_content));
+    }
+    ~StaticVariableWrapper()
+    {
+        DTOR::operator()(std::addressof(m_content));
+    }
+    T& Get() noexcept
+    {
+        return m_content;
+    }
+    const T& Get() const noexcept
+    {
+        return m_content;
+    }
+    operator T&() noexcept
+    {
+        return m_content;
+    }
+    operator const T&() const noexcept
+    {
+        return m_content;
+    }
+};
+/**
+ * @brief 生成静态变量包装类的函数
+ *
+ * @tparam T 要被包装的类型
+ * @tparam CTOR 自定义变量默认初始化后的函数类型
+ * @tparam DTOR 自定义变量执行析构函数前的函数类型
+ * @param ctor 自定义变量默认初始化后的行为，传入变量的指针作为参数
+ * @param dtor 自定义变量执行析构函数前的行为，传入变量的指针作为参数
+ * @return StaticVariableWrapper<T, DTOR> 包装后的变量，已经初始化
+ */
+template <class T, class CTOR, class DTOR = DefaultStaticVariableWrapperDtor<T>>
+auto MakeStaticVariableWrapper(CTOR ctor, DTOR dtor = {})
+    -> StaticVariableWrapper<T, DTOR>
+{
+    return {ctor, dtor};
+}
