+++ TrafficMonitor/D2D1Support.cpp
@@ -1,5 +1,6 @@
 ﻿#include "stdafx.h"
 #include "D2D1Support.h"
+#include "Common.h"
 #include "TrafficMonitor.h"
 
 #pragma comment(lib, "D2d1.lib")
